third_party/WebKit/LayoutTests/svg/animations/animateMotion-fill-freeze-expected.txt

SVG 1.1 dynamic animation tests

Test for animation freeze when repeatDur is not a multiple of dur

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is 0
PASS rootSVGElement.getBBox().x is 50
PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().x is 100
PASS successfullyParsed is true

TEST COMPLETE
